requirements.txt
pytest
selenium
webdriver-manager
python-dotenv
pytest-html
pytest-xdist
pytest-metadata
